+++ packages/contracts-bedrock/src/cannon/libraries/MIPSState.sol
@@ -3,9 +3,9 @@ pragma solidity 0.8.15;
 
 library MIPSState {
     struct CpuScalars {
-        uint32 pc;
-        uint32 nextPC;
-        uint32 lo;
-        uint32 hi;
+        uint64 pc;
+        uint64 nextPC;
+        uint64 lo;
+        uint64 hi;
     }
 }
